+++ include/QrCode.hpp
@@ -1,5 +1,5 @@
 /* 
- * QR Code generator library (C++)
+ * QR Code generator library (C++) - edited
  * 
  * Copyright (c) Project Nayuki. (MIT License)
  * https://www.nayuki.io/page/qr-code-generator-library
@@ -19,6 +19,9 @@
  *   liability, whether in an action of contract, tort or otherwise, arising from,
  *   out of or in connection with the Software or the use or other dealings in the
  *   Software.
+ * 
+ * Changes made:
+ *  - Add QrCode::getRawModules to get the raw module vectors
  */
 
 #pragma once
@@ -361,6 +364,8 @@ class QrCode final {
 	public: int getMask() const;
 	
 	
+	public: std::vector<std::vector<bool>> getRawModules() const;
+
 	/* 
 	 * Returns the color of the module (pixel) at the given coordinates, which is false
 	 * for light or true for dark. The top left corner has the coordinates (x=0, y=0).