+++ source/hooked_func.cpp
@@ -177,7 +177,7 @@ namespace CTRPluginFramework {
         cmdbuf[5] = pathData;
 
         Result ret = 0;
-        if (R_FAILED(ret = svcSendSyncRequest(*fsHandle))) customBreak(0xbad, ret, 0);
+        if (R_FAILED(ret = svcSendSyncRequest(*fsHandle))) customBreak(0xbad, 0xDEADFF1E, ret, 0);
 
         *out = cmdbuf[2] | ((u64)(cmdbuf[3]) << 32);
         DEBUG("fsOpenArch return 0x%08X%08X\n", cmdbuf[3], cmdbuf[2]);
@@ -185,13 +185,9 @@ namespace CTRPluginFramework {
         return cmdbuf[1];
     }
 
-    // Stubbed - SB3 doesn't need this besides the initial format.
+    // Stubbed
     int fsFormatSaveData(int *a1, int a2, int a3, int a4, int a5, int a6, int a7, int a8, int a9, int a10, char a11) {
         DEBUG("fsFormatSaveData called but the save data will not be removed.\n");
-        // std::string dir;
-        // Utils::ConvertUTF16ToUTF8(dir, OnionSave::dataPath);
-        // Directory::Remove(dir);
-        // Directory::Create(dir);
         return 0;
     }