+++ toCitra.sh
@@ -4,8 +4,6 @@ cp output/*.3gx /home/cyberyoshi64/.local/share/citra-emu/sdmc/luma/plugins/0004
 cp output/*.3gx /home/cyberyoshi64/.local/share/citra-emu/sdmc/luma/plugins/0004000000117200
 cp output/*.3gx /home/cyberyoshi64/.local/share/citra-emu/sdmc/luma/plugins/000400000016DE00
 cp output/*.3gx /home/cyberyoshi64/.local/share/citra-emu/sdmc/PTC3PLG/resources
-cp output/*.3gx /home/cyberyoshi64/.local/share/citra-emu/sdmc/PTC3PLG/resources
-cp output/*.3gx /home/cyberyoshi64/.local/share/citra-emu/sdmc/PTC3PLG/resources
 cp build/map_EUR.cyxmap /home/cyberyoshi64/.local/share/citra-emu/sdmc/PTC3PLG/resources
 cp build/map_USA.cyxmap /home/cyberyoshi64/.local/share/citra-emu/sdmc/PTC3PLG/resources
 cp build/map_JPN.cyxmap /home/cyberyoshi64/.local/share/citra-emu/sdmc/PTC3PLG/resources