+++ CHANGES.md
@@ -1,3 +1,9 @@
+# 0.0.7-0 (2024/3/5, 08:42)
+- Refactor some code
+- Fix Citra inconsistencies
+- Add BasicAPI experiments
+- Add save data check after plugin fail
+
 # 0.0.6-4 (2024/3/1, 23:22)
 
 - Replace `FwkSettings.WaitTimeToBoot` with `StallProcess()`